- Dodana JWT avtorizacija za klicanje knjig - spremenjen +page.server.js -> +page.js za dostop do spremenljivk okna
