--- BenjaminJeran/SOA_Bookcase_Frontend/src/routes/app/books/+page.server.js
@@ -1,15 +1,0 @@
-
-export async function load() {
-
-    const response = await fetch('http://localhost:3000/api/books');
-    
-    if (!response.ok) {
-        throw error(500, 'Failed to load books data');
-    }
-    
-    const books = await response.json();
-    
-    return {
-        books
-    };
-}
