--- BenjaminJeran/SOA_Bookcase_Frontend/src/routes/app/books/[id]/+page.server.js
@@ -1,21 +1,0 @@
-import { error } from '@sveltejs/kit';
-
-export async function load({ params }) {
- 
-    const { id } = params;
-    console.log(id);
-
-    try {
-        const response = await fetch(`http://localhost:3000/api/books/${id}/reviews`);
-
-        if (!response.ok) {
-            throw new Error(`Failed to fetch book with ID: ${id}`);
-        }
-
-        const book = await response.json();
-
-        return { props: { book } };
-    } catch (err) {
-        throw error(404, `Book with ID ${id} not found.`);
-    }
-}
